# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "QMDB - Movie Information API",
    "description": "This API provides access to publicly available data on a limited number of movies published from the year 1990 from the Internet Movie Database and other sources. The API endpoints and their usage are described in detail below. Modified by Zeyu Xia (n11398299) @ QUT 2023."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    },
    "schemas": {
...
      "MissingAuthHeader": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Authorization header ('Bearer token') not found"
          }
        }
      },
      "TokenExpired": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "JWT token has expired"
          }
        }
      },
      "InvalidJWT": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Invalid JWT token"
          }
        }
      },
...
    },
    "examples": {
...
      "MissingAuthHeader": {
        "summary": "MissingAuthHeader",
        "description": "When Authorization header ('Bearer token') is missing.",
        "value": {
          "error": "true",
          "message": "Authorization header ('Bearer token') not found"
        }
      },
      "TokenExpired": {
        "summary": "TokenExpired",
        "description": "If the JWT token has expired",
        "value": {
          "error": "true",
          "message": "JWT token has expired"
        }
      },
      "InvalidJWT": {
        "summary": "InvalidJWT",
        "description": "If the JWT token failed verification",
        "value": {
          "error": "true",
          "message": "Invalid JWT token"
        }
      },
...
    }
  },
  "paths": {
...
    "/people/{id}": {
      "get": {
        "tags": [
          "People"
        ],
        "description": "Get information about a person (actor, writer, director etc.) from their IMDB ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The person's IMDB ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "An object containing data about that person",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Patrick Stewart"
                    },
                    "birthYear": {
                      "type": "integer",
                      "example": 1940
                    },
                    "deathYear": {
                      "type": "integer",
                      "example": null
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "movieName": {
                            "type": "string",
                            "example": "Star Trek: First Contact"
                          },
                          "movieId": {
                            "type": "string",
                            "example": "tt0117731"
                          },
                          "category": {
                            "type": "string",
                            "example": "actor"
                          },
                          "characters": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "example": "Picard"
                            }
                          },
                          "imdbRating": {
                            "type": "number",
                            "example": 7.6
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters. Click on 'Schema' to see possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid query parameters: year. Query parameters are not permitted."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingAuthHeader"
                    },
                    {
                      "$ref": "#/components/schemas/TokenExpired"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidJWT"
                    }
                  ]
                },
                "examples": [
                  {
                    "$ref": "#/components/examples/MissingAuthHeader"
                  },
                  {
                    "$ref": "#/components/examples/TokenExpired"
                  },
                  {
                    "$ref": "#/components/examples/InvalidJWT"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "The requested person could not be found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "No record exists of a person with this ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}